Moved search bar component to utils
