+++ frontend/src/components/FAQHelp/FAQHelp.tsx
@@ -1,12 +1,12 @@
 import React, { ReactElement } from 'react';
-import HelpSearchBar from './HelpSearchBar';
+import SearchBar from '../utils/SearchBar';
 import styles from './FAQHelp.module.scss';
 
 const FAQHelp = (): ReactElement => {
   return (
     <div className={styles.helpComponent}>
       <h2 className={styles.helpTitle}>Need Help?</h2>
-      <HelpSearchBar />
+      <SearchBar placeholder="Search for any question..." ariaLabel="Search questions" />
     </div>
   );
 };
