--- frontend/src/components/FAQHelp/HelpSearchBar.module.scss
@@ -1,4 +1,0 @@
-.faqHelpSearch {
-  width: 60%;
-  margin: 10px auto 20px auto;
-}
