--- frontend/src/components/Home/SearchBar.module.scss
@@ -1,15 +1,0 @@
-.search {
-  width: 60%;
-  align-items: center;
-  margin-left: 20%;
-  margin-right: 20%;
-  margin-top: 10%;
-}
-
-@media only screen and (max-width: 600px) {
-  .search {
-    width: 90%;
-    margin-left: 5%;
-    margin-right: 5%;
-  }
-}
