+++ frontend/src/components/utils/SearchBar.module.scss
@@ -1,0 +1,4 @@
+.search {
+  width: 90%;
+  margin: 20px auto;
+}
