+++ frontend/src/components/utils/SearchBar.tsx
@@ -2,14 +2,19 @@ import React, { ReactElement } from 'react';
 import { ReactComponent as SearchIcon } from '../../assets/search.svg';
 import styles from './SearchBar.module.scss';
 
-const SearchBar = (): ReactElement => {
+type Props = {
+  readonly placeholder: string;
+  readonly ariaLabel: string;
+};
+
+const SearchBar = ({ placeholder, ariaLabel }: Props): ReactElement => {
   return (
     <div className={`${styles.search} input-group mb-3`}>
       <input
         type="text"
         className="form-control"
-        placeholder="Search by any location..."
-        aria-label="Search locations"
+        placeholder={placeholder}
+        aria-label={ariaLabel}
         aria-describedby="basic-addon2"
       />
       <div className="input-group-append">
