+++ frontend/src/pages/FAQPage.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement, useState, useEffect } from 'react';
 import FAQNavBar from '../components/FAQ/FAQNavBar';
-import FAQHelp from '../components/FAQHelp';
+import FAQHelp from '../components/FAQHelp/FAQHelp';
 import Faqs from '../components/FAQ';
 import ProgressSpinner from '../components/utils/ProgressSpinner';
 import axios from 'axios';
