+++ frontend/src/pages/HomePage.module.scss
@@ -6,3 +6,7 @@
   background-repeat: no-repeat;
   background-size: cover;
 }
+
+.search {
+  margin-top: 50px;
+}
