+++ frontend/src/pages/HomePage.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement } from 'react';
-import SearchBar from '../components/Home/SearchBar';
+import SearchBar from '../components/utils/SearchBar';
 import NavBar from '../components/Home/NavBar';
 import styles from './HomePage.module.scss';
 
@@ -7,7 +7,9 @@ const HomePage = (): ReactElement => {
   return (
     <div className={styles.Home}>
       <NavBar />
-      <SearchBar />
+      <div className={styles.search}>
+        <SearchBar placeholder="Search by any location..." ariaLabel="Search locations" />
+      </div>
     </div>
   );
 };
